requirements.txt
python-telegram-bot==21.5
Pillow==10.4.0
python-dotenv==1.0.1